Better defang code
